Cleanup controllers, mounted checks, deps, and UI alpha
Remove duplicate beneficiaryId TextEditingController initializations and redundant disposals; add mounted guards before Navigator.pop and ScaffoldMessenger calls to avoid use-after-dispose; remove unnecessary cast from clamp result; replace integer alpha values with fractional doubles (0.3 / 0.6) for consistent opacity; add the 'path' package to pubspec.yaml; convert RegExp consts to final and remove some file-level ignore comments. These are small cleanups and bugfixes to prevent runtime issues and improve code consistency.

+++ lib/src/core/repositories/dashboard_repository.dart
@@ -205,7 +205,7 @@ class DashboardRepository {
         );
 
         final remainingSlots =
-            (limit - activities.length).clamp(0, limit) as int;
+            (limit - activities.length).clamp(0, limit);
         for (final doc in disbursementDocs.take(remainingSlots)) {
           final data = doc.data();
           final amount = (data['reliefAmount'] as num?)?.toDouble() ?? 0.0;
